{"schema":"decodex/commit/1","summary":"Split scroll capture types and fingerprints","authority":"manual"}
